requirements.txt
flask==3.0.2
flask-cors==4.0.0
pdfplumber==0.11.0
PyPDF2==3.0.1
langchain==0.3.15
langchain-google-genai==2.0.8
google-generativeai==0.8.3
python-dotenv==1.0.1
requests==2.32.3
